+++ includes/checkout/thank-you.php
@@ -43,8 +43,10 @@ $order         = wc_get_order( $order_id );
 if ( $ty_is_rest ) {
 	if ( 'yes' === $ty_testmode ) {
 		if ( in_array(
-			strtoupper( $ty_klarna_country ), apply_filters(
-				'klarna_is_rest_countries_eu', array(
+			strtoupper( $ty_klarna_country ),
+			apply_filters(
+				'klarna_is_rest_countries_eu',
+				array(
 					'DK',
 					'GB',
 					'NL',
@@ -57,8 +59,10 @@ if ( $ty_is_rest ) {
 		}
 	} else {
 		if ( in_array(
-			strtoupper( $ty_klarna_country ), apply_filters(
-				'klarna_is_rest_countries_eu', array(
+			strtoupper( $ty_klarna_country ),
+			apply_filters(
+				'klarna_is_rest_countries_eu',
+				array(
 					'DK',
 					'GB',
 					'NL',
@@ -170,12 +174,15 @@ echo $snippet;
  * plugins like MonsterInsights pull data they need.
  */
 if ( true === apply_filters( 'klarna_finalize_order_in_thank_you_page', false ) ) {
-	$order->calculate_totals( false );
-	$order->update_status( 'pending' ); // Set status to Pending Payment before completing the order.
-	$order->payment_complete( $order_uri );
-	$order->set_date_created( current_time( 'timestamp', true ) );
-	delete_post_meta( klarna_wc_get_order_id( $order ), '_kco_incomplete_customer_email' );
-	add_post_meta( klarna_wc_get_order_id( $order ), '_kco_payment_created', time() );
+	// If the order already has been finalized in Woo, don't run payment_complete again.
+	if ( ! $order->has_status( array( 'processing', 'completed' ) ) ) {
+		$order->calculate_totals( false );
+		$order->update_status( 'pending' ); // Set status to Pending Payment before completing the order.
+		$order->payment_complete( $order_uri );
+		$order->set_date_created( current_time( 'timestamp', true ) );
+		delete_post_meta( klarna_wc_get_order_id( $order ), '_kco_incomplete_customer_email' );
+		add_post_meta( klarna_wc_get_order_id( $order ), '_kco_payment_created', time() );
+	}
 }
 
 do_action( 'klarna_after_kco_confirmation', $order_id, $klarna_order );
